set T to uint16

+++ svm_feature_gen_markovori.m
@@ -1,7 +1,7 @@
 function f=svm_feature_gen_markovori(autrain)
 % extract feature with original markov feature
 
-T=3;
+T=uint16(3);
 transmat=extract_feature(autrain,1);
 f=zeros(size(transmat,3),(2*T+1)^2);
